Refactor: Move ClampedPercentage type to common folder
