+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests/producer_gas_price_tests.rs
@@ -4,13 +4,14 @@ use fuel_core_gas_price_service::{
     static_updater::StaticAlgorithm,
 };
 use fuel_core_producer::block_producer::gas_price::GasPriceProvider;
+use fuel_core_types::clamped_percentage::ClampedPercentage;
 
 #[test]
 fn production_gas_price__if_requested_block_height_is_latest_return_gas_price() {
     // given
     let price = 33;
     let algo = StaticAlgorithm::new(price);
-    let gas_price_provider = build_provider(algo.clone(), 0, price, 10);
+    let gas_price_provider = build_provider(algo.clone(), 0, price, ClampedPercentage::new(10));
 
     // when
     let expected_price = algo.next_gas_price();
@@ -25,7 +26,7 @@ fn dry_run_gas_price__calculates_correctly_based_on_percentage() {
     // given
     let height = 123;
     let price = 33;
-    let percentage = 10;
+    let percentage = ClampedPercentage::new(10);
     let algo = StaticAlgorithm::new(price);
     let gas_price_provider = build_provider(algo.clone(), height, price, percentage);
 
@@ -33,7 +34,7 @@ fn dry_run_gas_price__calculates_correctly_based_on_percentage() {
     let actual = gas_price_provider.dry_run_gas_price().unwrap();
 
     // then
-    let change_amount = price.saturating_mul(percentage as u64).saturating_div(100);
+    let change_amount = price.saturating_mul(*percentage as u64).saturating_div(100);
     let expected = price + change_amount;
     assert_eq!(expected, actual);
 }
