+++ crates/fuel-core/src/service/sub_services.rs
@@ -60,6 +60,7 @@ use fuel_core_storage::{
 #[cfg(feature = "relayer")]
 use fuel_core_types::blockchain::primitives::DaBlockHeight;
 use fuel_core_types::signer::SignMode;
+use fuel_core_types::clamped_percentage::ClampedPercentage;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
@@ -222,9 +223,10 @@ pub fn init_sub_services(
         database.on_chain().clone(),
     )?;
     let (gas_price_algo, latest_gas_price) = gas_price_service_v1.shared.clone();
+    let clamped_percentage = ClampedPercentage::new(DEFAULT_GAS_PRICE_CHANGE_PERCENT as u8);
     let universal_gas_price_provider = UniversalGasPriceProvider::new_from_inner(
         latest_gas_price,
-        DEFAULT_GAS_PRICE_CHANGE_PERCENT,
+        clamped_percentage,
     );
 
     let producer_gas_price_provider = FuelGasPriceProvider::new(
