+++ crates/fuel-gas-price-algorithm/src/v1.rs
@@ -1,4 +1,5 @@
 use crate::utils::cumulative_percentage_change;
+use fuel_core_types::clamped_percentage::ClampedPercentage;
 use std::{
     cmp::{
         max,
@@ -318,36 +319,6 @@ impl L2ActivityTracker {
 
     pub fn block_activity_threshold(&self) -> ClampedPercentage {
         self.block_activity_threshold
-    }
-}
-
-/// A value that represents a value between 0 and 100. Higher values are clamped to 100
-#[derive(
-    serde::Serialize, serde::Deserialize, Debug, Copy, Clone, PartialEq, PartialOrd,
-)]
-pub struct ClampedPercentage {
-    value: u8,
-}
-
-impl ClampedPercentage {
-    pub fn new(maybe_value: u8) -> Self {
-        Self {
-            value: maybe_value.min(100),
-        }
-    }
-}
-
-impl From<u8> for ClampedPercentage {
-    fn from(value: u8) -> Self {
-        Self::new(value)
-    }
-}
-
-impl core::ops::Deref for ClampedPercentage {
-    type Target = u8;
-
-    fn deref(&self) -> &Self::Target {
-        &self.value
     }
 }
