+++ crates/types/src/lib.rs
@@ -32,6 +32,8 @@ pub mod blockchain;
 pub mod entities;
 pub mod services;
 pub mod signer;
+/// A module that provides functionality for working with clamped percentages.
+pub mod clamped_percentage;
 
 /// Re-export of some fuel-vm types
 pub mod fuel_vm {
